Collision checker function worked kind of counterintuitively
The collision checker function returned positive values when the point was not in collision. It seemed to me that a function for checking collisions would return positive values if collision is detected. Also, by doing so, the checking expression is simplified largely. Some additional variables for return values have been substituted for their corresponding True/False values direcly, for code simplicity.